+++ dev/children/vue2/src/App.vue
@@ -1,10 +1,16 @@
 <template>
   <div id="app">
-    <div class='tab-con'>
-      <el-tabs v-model="activeName" @tab-click="handleClick">
-        <el-tab-pane label="home" name="/"></el-tab-pane>
-        <el-tab-pane label="page2" name="page2"></el-tab-pane>
-      </el-tabs>
+    <div>
+      <el-menu
+        :default-active="activeIndex"
+        class="el-menu-demo"
+        mode="horizontal"
+        router
+      >
+        <el-menu-item index="/">home</el-menu-item>
+        <el-menu-item index="/page2">page2</el-menu-item>
+        <el-menu-item index="/table">table</el-menu-item>
+      </el-menu>
     </div>
     <div @click="reload">点击刷新</div>
     <!-- <keep-alive> -->
@@ -20,14 +26,16 @@ export default {
   name: 'App',
   data () {
     return {
-      activeName: location.href.includes('#/page2') ? 'page2': '/',
+      activeIndex: '',
       showView: true,
     }
   },
-  mounted () {
-    window.addEventListener('popstate', () => {
-      this.activeName =location.href.includes('#/page2') ? 'page2': '/'
+  created () {
+    this.$router.onReady(() => {
+      this.activeIndex = this.$route.path
     })
+  },
+  mounted () {
     document.getElementById('test-innerHTML').innerHTML = '<span>3333333333</span>'
   },
   components: {
@@ -42,6 +50,11 @@ export default {
       this.$nextTick(() => {
         this.showView = true
       })
+    }
+  },
+  watch: {
+    $route () {
+      this.activeIndex = this.$route.path
     }
   }
 }
@@ -58,6 +71,7 @@ export default {
   width: 100%;
   padding: 30px;
   box-sizing: border-box;
+  display: block;
 }
 
 .icon {
@@ -66,5 +80,9 @@ export default {
   vertical-align: -0.15em;
   fill: currentColor;
   overflow: hidden;
+}
+
+.el-menu-demo {
+  display: block;
 }
 </style>